Update jdb.tcl
Change the builtins given as args to unknown for renaming from a static list, to an algorithmically determined list. A change in the build for jimsh, such as changing extensions, then should not cause cryptic errors when certain builtins are not found. This latter problem causes an error: "append" cannot be renamed. (Which is cryptic--it relates to the process of error message construction for the absent builtin).